refactor(pytorch): package prefix cache trie
